pyproject.toml
[project]
name = "st-kmeans"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=5.5.0",
    "folium>=0.19.5",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pystac-client>=0.8.6",
    "rasterio>=1.4.3",
    "scikit-learn>=1.6.1",
    "streamlit>=1.44.1",
    "streamlit-folium>=0.24.0",
    "loguru>=0.7.3",
    "dotenv>=0.9.9",
    "shapely>=2.1.0",
    "upython>=0.2",
    "ipython>=9.0.2",
]
